feat : 클라이언트에서 서버로 로그인 / 회원가입 요청/응답 처리 추가
(언리얼 클라이언트 <-> 로그인 서버 통신 처리 확인)

+++ TCPLoginServer/DB/DBQueryStatus.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 
-enum class SignUpStatus : uint16_t
+enum class SignUpStatus : uint8_t
 {
 	Success,
 	DuplicateUserId,
@@ -9,7 +9,7 @@ enum class SignUpStatus : uint16_t
 	OtherError
 };
 
-enum class LoginStatus : uint16_t
+enum class LoginStatus : uint8_t
 {
 	Success,
 	UserIdNotFound,
